requirements.txt
PyJWT===1.7.1
django

